管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -42,6 +42,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
+
         //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -63,7 +66,7 @@ public class EmployeeAction extends ActionBase {
 
         //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
-
+        }
     }
 
     /**
@@ -73,11 +76,15 @@ public class EmployeeAction extends ActionBase {
      */
     public void entryNew() throws ServletException, IOException {
 
+        //管理者かどうかチェック
+        if(checkAdmin()) {
+
         putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
         //新規登録画面を表示
         forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /**
@@ -87,8 +94,8 @@ public class EmployeeAction extends ActionBase {
      */
     public void create() throws ServletException, IOException {
 
-        //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        //CSRF対策 tokenのチェック と管理者かどうかチェック
+        if (checkToken() && checkAdmin()) {
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -136,6 +143,7 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void show() throws ServletException, IOException {
+        if(checkAdmin()) {
 
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -151,6 +159,7 @@ public class EmployeeAction extends ActionBase {
 
         //詳細画面を表示
         forward(ForwardConst.FW_EMP_SHOW);
+        }
     }
 
     /*
@@ -159,6 +168,7 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void edit() throws ServletException, IOException{
+        if(checkAdmin()) {
         //idを条件に従業員データを取得
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -172,6 +182,7 @@ public class EmployeeAction extends ActionBase {
 
         //編集画面の表示
         forward(ForwardConst.FW_EMP_EDIT);
+        }
     }
 
     /*
@@ -180,8 +191,8 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void update() throws ServletException,IOException{
-        //CSRF対策
-        if(checkToken()) {
+        //CSRF対策 と 管理者かどうかチェック
+        if(checkToken() && checkAdmin()) {
             //パラメータの値をもとに従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -207,7 +218,7 @@ public class EmployeeAction extends ActionBase {
                 //編集画面再表示
                 forward(ForwardConst.FW_EMP_EDIT);
             }else {
-                //更新中にエラーがなっかた場合
+                //更新中にエラーがなかった場合
 
                 //セッションに更新完了のフラッシュメッセージを設定
                 putSessionScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
@@ -225,8 +236,8 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void destroy() throws ServletException, IOException{
-        //CSRF対策 tokenチェック
-        if(checkToken()) {
+        //CSRF対策 tokenチェック と 管理者かどうかチェック
+        if(checkToken()&& checkAdmin()) {
             //idを条件に従業員データを論理削除
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -235,6 +246,25 @@ public class EmployeeAction extends ActionBase {
 
             //一覧画面にリダイレクト
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
+    }
+
+    /*
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * @true :管理者 false: 管理者でない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException{
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面表示
+        if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        }else {
+            return true;
         }
     }
